refactor : renamed CardOptions to TaskOptions
